Backend2/requirements.txt
python-dotenv
psycopg2-binary
pgvector
openai
fastapi
pydantic
google-generativeai
google-adk
requests
uvicorn
